+++ src/src/entity.cpp
@@ -6,7 +6,7 @@
 #include <scene.hpp>
 #include <Instrumentor.h>
 
-uint64_t NEXT_UID=0;
+uint32_t NEXT_UID=0;
 
 Entity::Entity(): m_X(0.0f),m_Y(0.0f),m_PreviousX(0.0f),m_PreviousY(0.0f),m_UID(NEXT_UID++){}
 Entity::Entity(float x,float y): m_X(x),m_Y(y),m_PreviousX(x),m_PreviousY(y),m_UID(NEXT_UID++){}
@@ -45,16 +45,16 @@ Entity &Entity::operator=(Entity &&other){
     return *this;
 }
 
-TextureComponent::TextureComponent(const std::string &path,int mag_filter,int min_filter,float width,float height,float layer,uint64_t uid):
+TextureComponent::TextureComponent(const std::string &path,int mag_filter,int min_filter,float width,float height,float layer,uint32_t uid):
     m_Texture(std::make_shared<Texture>(path,mag_filter,min_filter)),m_Width(width),m_Height(height),m_Layer(layer),m_UID(uid){}
 
-TextureComponent::TextureComponent(std::shared_ptr<Texture>&t,float width,float height,float layer,uint64_t uid):
+TextureComponent::TextureComponent(std::shared_ptr<Texture>&t,float width,float height,float layer,uint32_t uid):
     m_Texture(t),m_Width(width),m_Height(height),m_Layer(layer),m_UID(uid){}
 
-AnimatedTextureComponent::AnimatedTextureComponent(const std::string &path,unsigned int tile_width,unsigned int tile_height,int mag_filter,int min_filter,float width,float height,float layer,uint64_t uid):
+AnimatedTextureComponent::AnimatedTextureComponent(const std::string &path,unsigned int tile_width,unsigned int tile_height,int mag_filter,int min_filter,float width,float height,float layer,uint32_t uid):
     m_AnimatedTexture(std::make_shared<AnimatedTexture>(path,tile_width,tile_height,mag_filter,min_filter)),m_Width(width),m_Height(height),m_Layer(layer),m_UID(uid){}
 
-AnimatedTextureComponent::AnimatedTextureComponent(std::shared_ptr<AnimatedTexture>&t,float width,float height,float layer,uint64_t uid):
+AnimatedTextureComponent::AnimatedTextureComponent(std::shared_ptr<AnimatedTexture>&t,float width,float height,float layer,uint32_t uid):
     m_AnimatedTexture(t),m_Width(width),m_Height(height),m_Layer(layer),m_UID(uid){}
 
 TextureComponent::TextureComponent(TextureComponent &other){
@@ -89,8 +89,8 @@ void AnimatedTextureComponent::PlayAnimation(bool loop,float delay){
     m_AnimatedTexture->PlayAnimation(loop,delay);
 }
 
-LightComponent::LightComponent(): m_XOffset(0.0f),m_YOffset(0.0f),m_Radius(0.0f),m_Blur(0.0f),m_Color(0.0f,0.0f,0.0f),m_Type(LIGHT_AROUND_POS),m_UID(std::numeric_limits<uint64_t>::max()){}
-LightComponent::LightComponent(float x_offset,float y_offset,float radius,float blur,Vec3 color,LightType type,uint64_t uid): 
+LightComponent::LightComponent(): m_XOffset(0.0f),m_YOffset(0.0f),m_Radius(0.0f),m_Blur(0.0f),m_Color(0.0f,0.0f,0.0f),m_Type(LIGHT_AROUND_POS),m_UID(std::numeric_limits<uint32_t>::max()){}
+LightComponent::LightComponent(float x_offset,float y_offset,float radius,float blur,Vec3 color,LightType type,uint32_t uid): 
     m_XOffset(x_offset),m_YOffset(y_offset),m_Radius(radius),m_Blur(blur),m_Color(color),m_Type(type),m_UID(uid){}
 
 void LightComponent::SetOffset(float x_offset,float y_offset){
